+++ Tests/TuistGeneratorTests/Generator/Mocks/MockWorkspaceGenerator.swift
@@ -3,16 +3,22 @@ import Foundation
 import TuistCore
 import TuistCoreTesting
 import TuistSupport
+import XcodeProj
 @testable import TuistGenerator
 
 final class MockWorkspaceGenerator: WorkspaceGenerating {
-    var generateWorkspaces: [Workspace] = []
-    var generateStub: ((Workspace, AbsolutePath, Graphing) throws -> AbsolutePath)?
+    enum MockError: Error {
+        case stubNotImplemented
+    }
 
-    func generate(workspace: Workspace,
-                  path: AbsolutePath,
-                  graph: Graphing) throws -> AbsolutePath {
+    var generateWorkspaces: [Workspace] = []
+    var generateStub: ((Workspace, AbsolutePath, Graphing) throws -> WorkspaceDescriptor)?
+
+    func generate(workspace: Workspace, path: AbsolutePath, graph: Graphing) throws -> WorkspaceDescriptor {
+        guard let generateStub = generateStub else {
+            throw MockError.stubNotImplemented
+        }
         generateWorkspaces.append(workspace)
-        return (try generateStub?(workspace, path, graph)) ?? AbsolutePath("/test")
+        return try generateStub(workspace, path, graph)
     }
 }